requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai>=1.3.8,<2
pillow==10.1.0
pytesseract==0.3.10
aiofiles==23.2.0
slowapi==0.1.9
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
structlog==23.2.0
prometheus-client==0.19.0
PyPDF2==3.0.1
pdfplumber>=0.11.7
pdf2markdown4llm>=0.1.0
PyMuPDF>=1.23.0
pdf2image>=1.17.0
pydub==0.25.1
duckduckgo-search>=6.0.0
python-jose[cryptography]>=3.3.0
google-auth>=2.23.0
requests>=2.31.0
SQLAlchemy>=2.0.0
PyMySQL>=1.1.0
boto3>=1.28.0
celery>=5.3.0
psycopg2-binary>=2.9.9
pgvector>=0.3.0
tiktoken>=0.5.0
flashrank>=0.2.0
